Factor the NVMe types out of unvme_nvme.h
The NVMe types are valuable on their own, but the header file cannot
be reused out of context if it includes function prototypes.
